Fix CG022 binding: correct CRC polynomial (0x8005), CRC init (0x4402), preamble (3 bytes)
Three discrepancies between NRF24L01 emulated output and original LT8910 SPI captures:

1. CRC polynomial: emulation used default 0x1021 (CRC-CCITT) but LT8910
   register 0x17 = 0x8005 (CRC-16/USB). Wrong CRC = all packets rejected.

2. CRC init: emulation used 0x0000 but register 0x28 = 0x4402 (16-bit seed).
   Expanded LT8900_CRC_Initial_Data from uint8_t to uint16_t. SHENQI
   backward compatible (0xAA promotes to 0x00AA).

3. Preamble: register 0x20 bits [15:13] = 010 → preamble = 3 bytes
   (field value + 1). Was incorrectly set to 2.

+++ Multiprotocol/NRF24l01_SPI.ino
@@ -253,7 +253,7 @@ uint8_t LT8900_addr[8];
 uint8_t LT8900_addr_size;
 uint8_t LT8900_Preamble_Len;
 uint8_t LT8900_Tailer_Len;
-uint8_t LT8900_CRC_Initial_Data;
+uint16_t LT8900_CRC_Initial_Data;
 uint8_t LT8900_Flags;
 #define LT8900_CRC_ON 6
 #define LT8900_SCRAMBLE_ON 5
@@ -263,7 +263,7 @@ uint8_t LT8900_Flags;
 #define LT8900_FEC_TYPE_1 1
 #define LT8900_FEC_TYPE_0 0
 
-void LT8900_Config(uint8_t preamble_len, uint8_t trailer_len, uint8_t flags, uint8_t crc_init)
+void LT8900_Config(uint8_t preamble_len, uint8_t trailer_len, uint8_t flags, uint16_t crc_init)
 {
 	//Preamble 1 to 8 bytes
 	LT8900_Preamble_Len=preamble_len;
